+++ docs/roi2mask.md
@@ -15,10 +15,6 @@ Convert a region of interest/object contour to a binary mask of the same shape
 - **Example use:**
     - below
 
-**Original RGB image**
-
-![Screenshot](img/documentation_images/rgb2lab/original_image.jpg)
-
 ```python
 
 from plantcv import plantcv as pcv
@@ -38,42 +34,8 @@ mask = pcv.roi2mask(img=img, contour=roi_contour)
 
 ```
 
-**Lightness channel image**
+**Custom ROI Image**
 
-![Screenshot](img/documentation_images/rgb2lab/lab_lightness.jpg)
+![Screenshot](img/documentation_images/roi2mask/lab_lightness.jpg)
 
-```python
 
-from plantcv import plantcv as pcv
-
-# Set global debug behavior to None (default), "print" (to file), 
-# or "plot" (Jupyter Notebooks or X11)
-pcv.params.debug = "print"
-
-# image converted from RGB to LAB, channels are then split. 
-# Green-Magenta ('a') channel is outputed.
-a_channel = pcv.rgb2gray_lab(rgb_img, 'a')
-
-```
-
-**Green-Magenta channel image**
-
-![Screenshot](img/documentation_images/rgb2lab/lab_green-magenta.jpg)
-   
-```python
-
-from plantcv import plantcv as pcv
-
-# Set global debug behavior to None (default), "print" (to file), 
-# or "plot" (Jupyter Notebooks or X11)
-pcv.params.debug = "print"
-
-# image converted from RGB to Lab, channels are then split. 
-# Blue-Yellow ('b') channel is outputed.
-b_channel = pcv.rgb2gray_lab(rgb_img, 'b')
-
-```
-
-**Blue-Yellow channel image**
-
-![Screenshot](img/documentation_images/rgb2lab/lab_blue-yellow.jpg)
